+++ litert/vendors/qualcomm/debugger/assets/config.json
