+++ litert/vendors/qualcomm/debugger/assets/config_viewer.json
