+++ litert/vendors/qualcomm/debugger/assets/htp_ext_config.json
