+++ litert/vendors/qualcomm/debugger/assets/qhas.png
